+++ app/src/main/java/com/example/verseverwebt/valueobjects/permissions/PermissionManager.kt
@@ -1,4 +1,4 @@
-package com.example.verseverwebt.user.permissions
+package com.example.verseverwebt.valueobjects.permissions
 
 import android.Manifest
 import android.app.Activity
